// frontend/src/models/enums/speciality.models.ts
export enum SpecialityEnum {
    ANESTHESIOLOGY = "ANESTHESIOLOGY",
    CARDIOLOGY = "CARDIOLOGY",
    GENERAL_SURGERY = "GENERAL_SURGERY",
    PLASTIC_SURGERY = "PLASTIC_SURGERY",
    DERMATOLOGY = "DERMATOLOGY",
    ENDOCRINOLOGY = "ENDOCRINOLOGY",
    GASTROENTEROLOGY = "GASTROENTEROLOGY",
    GERIATRICS = "GERIATRICS",
    GYNECOLOGY_AND_OBSTETRICS = "GYNECOLOGY_AND_OBSTETRICS",
    HEMATOLOGY = "HEMATOLOGY",
    INFECTIOUS_DISEASE = "INFECTIOUS_DISEASE",
    FAMILY_AND_GENERAL_MEDICINE = "FAMILY_AND_GENERAL_MEDICINE",
    INTERNAL_MEDICINE = "INTERNAL_MEDICINE",
    NEPHROLOGY = "NEPHROLOGY",
    PULMONOLOGY = "PULMONOLOGY",
    NEUROSURGERY = "NEUROSURGERY",
    NEUROLOGY = "NEUROLOGY",
    NUTRITION = "NUTRITION",
    OPHTHALMOLOGY = "OPHTHALMOLOGY",
    ONCOLOGY = "ONCOLOGY",
    ORTHOPEDICS_AND_TRAUMATOLOGY = "ORTHOPEDICS_AND_TRAUMATOLOGY",
    OTORHINOLARYNGOLOGY = "OTORHINOLARYNGOLOGY",
    PEDIATRICS = "PEDIATRICS",
    PSYCHIATRY = "PSYCHIATRY",
    RADIOLOGY = "RADIOLOGY",
    RHEUMATOLOGY = "RHEUMATOLOGY",
    UROLOGY = "UROLOGY"
}